// apps/auth/src/account/api/representations.ts
export interface Client {
  baseUrl: string;
  clientId: string;
  name?: string;
}

export interface Scope {
  name: string;
  displayName?: string;
}

export interface Resource {
  _id: string;
  name: string;
  client: Client;
  scopes: Scope[];
  uris: string[];
  shareRequests?: Permission[];
}

export interface Permission {
  email?: string;
  firstName?: string;
  lastName?: string;
  scopes: Scope[] | string[]; // this should be Scope[] - fix API
  username: string;
}